use parseFloat over parseInt

+++ src/routes/models/scheme-payments-by-year.js
@@ -52,7 +52,7 @@ function transformSummary (schemePaymentsByYear) {
   Object.keys(schemePaymentsByYear).forEach(year => {
     const formattedYear = getFormattedYear(year)
     schemePaymentsSummary[formattedYear] = schemePaymentsByYear[year].map(scheme => ({
-      ...scheme, total_amount: getReadableAmount(parseInt(scheme.total_amount))
+      ...scheme, total_amount: getReadableAmount(parseFloat(scheme.total_amount))
     }))
   })
